refactor: UniversityService, UserService 인터페이스 제거
