+++ server/momo-core/src/main/java/com/momo/user/application/UniversityService.java
@@ -1,9 +1,22 @@
 package com.momo.user.application;
 
+import com.momo.user.application.dto.UniversityDtoAssembler;
 import com.momo.user.application.dto.response.UniversityResponseDto;
+import com.momo.user.domain.dto.UniversityDto;
+import com.momo.user.domain.repository.UniversitySearchRepository;
 import java.util.List;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
 
-public interface UniversityService {
+@Service
+@RequiredArgsConstructor
+public class UniversityService {
 
-    List<UniversityResponseDto> findUniversity(String universityName);
+    private final UniversitySearchRepository universitySearchRepository;
+
+    public List<UniversityResponseDto> findUniversities(String universityName) {
+        List<UniversityDto> universityDtos = universitySearchRepository.searchUniversitiesByName(universityName);
+
+        return UniversityDtoAssembler.mapToUniversityResponseDtos(universityDtos);
+    }
 }
